+++ pgdrive/scene_creator/vehicle_module/lidar.py
@@ -1,11 +1,10 @@
 import logging
-from pgdrive.pg_config.collision_group import CollisionGroup
 from typing import Set
 
 import numpy as np
 from panda3d.bullet import BulletGhostNode, BulletSphereShape
 from panda3d.core import BitMask32, NodePath
-from pgdrive.constants import BodyName, CamMask
+from pgdrive.constants import BodyName, CamMask, CollisionGroup
 from pgdrive.scene_creator.highway_vehicle.behavior import IDMVehicle
 from pgdrive.scene_creator.vehicle.traffic_vehicle import PGTrafficVehicle
 from pgdrive.utils.asset_loader import AssetLoader
