+++ pgdrive/world/terrain.py
@@ -2,8 +2,7 @@ import numpy
 from panda3d.bullet import BulletRigidBodyNode, BulletPlaneShape
 from panda3d.core import Vec3, CardMaker, LQuaternionf, BitMask32, NodePath, TextureStage, Texture, SamplerState
 
-from pgdrive.constants import BodyName, CamMask
-from pgdrive.pg_config.collision_group import CollisionGroup
+from pgdrive.constants import BodyName, CamMask, CollisionGroup
 from pgdrive.utils.asset_loader import AssetLoader
 from pgdrive.utils.element import Element
